Minor changes to allow editing of new text recipes, converting to PBR overlays
There are two main changes in this changeset. The first change it to
allow editing the RaceData on a recipe. A newly created recipe doesn’t
have a RaceData, and the editor will fail without a race.
The second change is in the SlotData. This change is to make it easier
to deal with recipes that have saved ColorDatas, when changing the
overlay from the Legacy UMAMaterial to a PBR Material (it has more
channels). This change will expand the color data to match an incoming
material by applying default values to the new channels.