Significant speed increase when "fast generation" is turned on and converting render textures.
When converting RenderTextures to regular textures (to save texture memory, and compress textures), this gives a significant speed improvement by basically removing the yields, and processing each texture as a whole instead of pieceemeal.  In the crowd scene, this improved response time from 6200 ms for 25 UMAS to 2700 ms for 25 UMAS. Note: this works off the UMAGenerator "Fast Generation" flag that already exists.