Added specific culture settings for obj strings
When exporting an avatar as obj file using this class, in some machines the float values for vertices, normals and uv could end up written with a comma (",") as separator instead of a period ("."). That would make the resulting obj file not readable by most editor/viewer applications. This fix solves that issues.